requirements.txt
sentence_transformers
numpy
pandas
matplotlib
newspaper3k
scikit-learn
regex
joblib
plotly
argparse
tqdm